Move SAML nested types to internal package.
